backend/requirements.txt
amqp==5.1.1
APScheduler==3.9.1.post1
asgiref==3.5.2
astropy==5.1.1
astroquery==0.4.6
attrs==23.1.0
beautifulsoup4==4.12.1
billiard==3.6.4.0
black==24.1.1
cached-property==1.5.2
Cartopy==0.22.0
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
colorlog==6.8.2
contourpy==1.0.6
coverage==7.4.1
cron-descriptor==1.4.0
cryptography==41.0.4
cycler==0.11.0
Django==3.2.18
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==3.11.0
django-filter==21.1
django-settings-export==1.2.1
django-shibboleth-remoteuser==0.12
django-timezone-field==5.0
django-url-filter==0.3.15
djangorestframework==3.14.0
enum-compat==0.0.3
Fiona==1.9.4.post1
flower==1.2.0
fonttools==4.38.0
greenlet==2.0.1
html5lib==1.1
humanize==4.4.0
idna==3.4
importlib-metadata==6.1.0
isort==5.12.0
jaraco.classes==3.2.3
jeepney==0.8.0
kaleido==0.2.1
keyring==23.13.1
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.1
matplotlib==3.6.2
more-itertools==9.1.0
numpy==1.23.4
OWSLib==0.29.2
packaging
pandas==1.5.1
Pillow==9.3.0
plotly==5.14.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
pyerfa==2.0.0.1
pykdtree==1.3.9
pyparsing==3.0.9
pyproj==3.6.1
pyshp==2.3.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyvo==1.4.1
PyYAML==6.0
requests==2.28.1
scipy==1.9.3
SecretStorage==3.3.3
shapely==2.0.1
six==1.16.0
sora-astro==0.3
soupsieve==2.4
spiceypy==5.1.2
SQLAlchemy==1.4.44
sqlparse==0.4.3
tenacity==8.2.2
tornado==6.3.3
tqdm==4.66.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
uWSGI==2.0.22
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
zipp==3.15.0
